requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.4
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.3.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==12.0.1
iniconfig==1.1.1
kombu==5.2.4
librabbitmq==2.0.0
mixer==7.2.2
msgpack==1.0.4
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
